Add files via upload
Master in architecture with specialization in Theory and Practice of the Architectural Project, Technical University of Catalonia. In Chile, Spain and The Netherlands he has exercised as an Architect and an Interior Designer for a large list of projects.

His academic and professional background in architecture and interior design has led to a wide range of experience developing projects for private, commercial clients and for local authorities. His projects over the years have included public space regeneration, large residential development and interior design – all with focus on creating places in which people can lead better life.

Hector has teaching experience in pre-degree university studies and in development of interdisciplinary university workshops related to architecture and photography. 

Currently works for Cravt in Amsterdam as an Interior Architect and Design Director, leading the entire project development process.